refactor: allow more client customization
