+++ pyproject.toml
@@ -9,10 +9,10 @@ dependencies = [
     "fastapi>=0.115.8",
     "geojson-pydantic>=1.2.0",
     "pydantic>=2.10.4",
+    "rustac",
     "stac-fastapi-api>=5.0.2",
     "stac-fastapi-extensions>=5.0.2",
     "stac-fastapi-types>=5.0.2",
-    "rustac==0.6.0",
 ]
 classifiers = [
     "Programming Language :: Python :: 3",
@@ -58,7 +58,10 @@ deploy = [
 ]
 
 [tool.uv]
-default-groups = ["dev", "validate", "deploy"]
+default-groups = ["dev", "validate"]
+
+[tool.uv.sources]
+rustac = { git = "https://github.com/stac-utils/rustac-py", rev = "22c883178cb799032c0928eeaf82463b38f6bd7a" }
 
 [tool.mypy]
 strict = true
