+++ tests/conftest.py
@@ -5,16 +5,16 @@ import pytest
 from fastapi.testclient import TestClient
 from pytest import FixtureRequest
 
-import stac_fastapi.geoparquet
+import stac_fastapi.geoparquet.api
 from stac_fastapi.geoparquet import Settings
 
-geoparquet_file = Path(__file__).parents[1] / "data" / "naip.parquet"
-toml_file = Path(__file__).parents[1] / "data" / "config.toml"
+GEOPARQUET_FILE = Path(__file__).parents[1] / "data" / "naip.parquet"
+TOML_FILE = Path(__file__).parents[1] / "data" / "config.toml"
 
 
-@pytest.fixture(params=[geoparquet_file, toml_file])
+@pytest.fixture(params=[GEOPARQUET_FILE, TOML_FILE])
 async def client(request: FixtureRequest) -> AsyncIterator[TestClient]:
     settings = Settings(stac_fastapi_geoparquet_href=str(request.param))
-    api = stac_fastapi.geoparquet.create_api(settings)
+    api = stac_fastapi.geoparquet.api.create(settings)
     with TestClient(api.app) as client:
         yield client
